+++ src/helpers/SiteUriHelper.php
@@ -120,9 +120,15 @@ class SiteUriHelper
                 'uri' => $uri,
             ]);
 
-            if (!$cacheableOnly || Blitz::$plugin->cacheRequest->getIsCacheableSiteUri($siteUri)) {
-                $siteUris[] = $siteUri;
+            if ($cacheableOnly) {
+                if (!Blitz::$plugin->cacheRequest->getIsCacheableSiteUri($siteUri)
+                    || !Blitz::$plugin->settings->generatePageBasedOnQueryString($uri)
+                ) {
+                    continue;
+                }
             }
+
+            $siteUris[] = $siteUri;
         }
 
         return $siteUris;